Day 136: Tag-based Asset Retrieval

+++ code/handmade_sim_region.h
@@ -80,7 +80,7 @@ struct sim_entity {
 
     sim_entity_collision_volume_group *Collision;
 
-    uint32 FacingDirection;
+    real32 FacingDirection;
     real32 tBob;
 
     int32 dAbsTileZ;
